Add auth_options resolution from config (#2)
New feature to allow passing different client options while authenticating.
Previously only allowed client_options to be raw passed to the wrapped package. There are occations when for example an auth service resides on the internet and the web service resides in an internal network, where it is needed to only use proxy for the auth but not the call.

We now introduced "auth_options" configuration to allow the same methodology of passing raw options to the wrapped client independently.